Add table of plan
