+++ src/components/Course/index.js
@@ -1,14 +1,13 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { Tabs, TabList, Tab, TabPanel } from '@zendeskgarden/react-tabs';
+import { Tabs, TabList, Tab } from '@zendeskgarden/react-tabs';
 import { ThemeProvider } from '@zendeskgarden/react-theming';
 import { Introduction } from './Introduction';
 import { CourseClass } from './CourseClass';
 import { Resourses } from './Resourses';
 import { Practice } from './Practice';
-import { ContainerTitle } from '../../components/UI/Content';
+import { Plan } from './Plan';
 
-import { Table } from '../../components/Tables';
 // eslint-disable-next-line
 const TABS = [
   {
@@ -50,60 +49,6 @@ const StyledTabsList = styled(TabList)`
   justify-content: space-around;
 `;
 
-const Plan = () => {
-  const columns = React.useMemo(
-    () => [
-      {
-        Header: 'Tema',
-        accessor: 'theme',
-      },
-      {
-        Header: 'Objetivo',
-        accessor: 'objective',
-      },
-      {
-        Header: 'Actividades',
-        accessor: 'activitie',
-      },
-      {
-        Header: 'Materiales',
-        accessor: 'material',
-      },
-      {
-        Header: 'Min.Clase',
-        accessor: 'class',
-      },
-    ],
-    []
-  );
-  const data = React.useMemo(
-    () => [
-      {
-        theme: 'Alejandro',
-        objective: 'Medina',
-        activitie: 'relationship',
-        material: 'lalalala',
-        class: Math.floor(Math.random() * 100),
-      },
-      {
-        theme: 'Alejandro',
-        objective: 'Medina',
-        activitie: 'relationship',
-        material: 'lalalala',
-        class: Math.floor(Math.random() * 100),
-      },
-    ],
-    []
-  );
-  return (
-    <TabPanel item='tab-5'>
-      <ContainerTitle size='2' align='center' color='var(--color-purple)'>
-        ESTRUCTURA DE LA CLASE
-      </ContainerTitle>
-      <Table columns={columns} data={data} color='var(--color-purple)' />
-    </TabPanel>
-  );
-};
 export const Course = () => {
   const [selectedTab, setSelectedTab] = useState('tab-1');
